Move ray_config.h
