+++ src/ray/common/client_connection.cc
@@ -4,7 +4,7 @@
 #include <boost/bind.hpp>
 #include <sstream>
 
-#include "ray/ray_config.h"
+#include "ray/common/ray_config.h"
 #include "ray/raylet/format/node_manager_generated.h"
 #include "ray/util/util.h"
