+++ src/ray/gcs/client_test.cc
@@ -7,7 +7,7 @@ extern "C" {
 
 #include "ray/gcs/client.h"
 #include "ray/gcs/tables.h"
-#include "ray/ray_config.h"
+#include "ray/common/ray_config.h"
 
 namespace ray {
