+++ src/ray/raylet/monitor.cc
@@ -1,6 +1,6 @@
 #include "ray/raylet/monitor.h"
 
-#include "ray/ray_config.h"
+#include "ray/common/ray_config.h"
 #include "ray/common/status.h"
 #include "ray/util/util.h"
