+++ src/ray/raylet/monitor_main.cc
@@ -1,6 +1,6 @@
 #include <iostream>
 
-#include "ray/ray_config.h"
+#include "ray/common/ray_config.h"
 #include "ray/raylet/monitor.h"
 #include "ray/util/util.h"
