+++ src/ray/raylet/raylet_client.cc
@@ -13,7 +13,7 @@
 #include <sys/un.h>
 
 #include "ray/common/common_protocol.h"
-#include "ray/ray_config.h"
+#include "ray/common/ray_config.h"
 #include "ray/raylet/format/node_manager_generated.h"
 #include "ray/raylet/task_spec.h"
 #include "ray/util/logging.h"
